requirements.txt
dvc==3.50.1
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
dvclive==3.46.0
scikit-image==0.22.0
scikit-learn==1.3.1
scikit-optimize==0.9.0
scipy==1.11.1
numpy==1.26.2
flask
pytest
redis
python-dotenv==1.0.1